Convert references navigation tests to use jest
