+++ modules/references/client/components/create-reference/Navigation.js
@@ -11,6 +11,7 @@ export function Navigation({ t, disabled, tab, tabs, tabDone, onBack, onNext, on
   const backButton = (
     <button
       type="button"
+      role="button"
       className="btn btn-action btn-link"
       aria-label="Previous section"
       onClick={onBack}>
@@ -22,6 +23,7 @@ export function Navigation({ t, disabled, tab, tabs, tabDone, onBack, onNext, on
   const nextButton = (
     <button
       type="button"
+      role="button"
       className="btn btn-action btn-primary"
       aria-label="Next section"
       onClick={onNext}
@@ -32,6 +34,7 @@ export function Navigation({ t, disabled, tab, tabs, tabDone, onBack, onNext, on
 
   const submitButton = (
     <button
+      role="button"
       className="btn btn-action btn-primary"
       aria-label="Submit reference"
       onClick={onSubmit}
